Integration tests for master when deltas are enabled. Make sure we don't try to extract a file that doesn't exist.

+++ lib/blockbuster/concerns/extractor.rb
@@ -2,6 +2,7 @@ module Blockbuster
   # extracts files from gzipped tarballs
   module Extractor
     def extract_cassettes
+      return unless File.exist?(file_path)
       File.open(file_path, 'rb') do |file|
         Zlib::GzipReader.wrap(file) do |gz|
           Gem::Package::TarReader.new(gz) do |tar|
